feat: change auth module to CQRS pattern
