+++ src/auth/auth.module.ts
@@ -1,19 +1,22 @@
 import { TypeOrmModule } from '@nestjs/typeorm';
-
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { User } from './entities/user.entity';
-import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
+import { PassportModule } from '@nestjs/passport';
 import { JwtStrategy } from './strategies/jwt.strategy';
+import { CqrsModule } from '@nestjs/cqrs';
 
+import { AuthController } from './auth.controller';
+import { User } from './entities/user.entity';
+import { CreateUserHandler, LoginUserHandler } from './commands/handlers';
+
+export const CommandHandlers = [CreateUserHandler, LoginUserHandler];
 @Module({
   controllers: [AuthController],
-  providers: [AuthService, JwtStrategy],
+  providers: [JwtStrategy, ...CommandHandlers],
   imports: [
     ConfigModule,
+    CqrsModule,
     TypeOrmModule.forFeature([User]),
     PassportModule.register({ defaultStrategy: 'jwt' }),
     JwtModule.registerAsync({
