--- src/auth/auth.service.ts
@@ -1,77 +1,0 @@
-import {
-  BadRequestException,
-  Injectable,
-  InternalServerErrorException,
-  Logger,
-  UnauthorizedException,
-} from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import * as bcrypt from 'bcrypt';
-import { JwtService } from '@nestjs/jwt';
-import { Repository } from 'typeorm';
-
-import { LoginUserDto } from './dto';
-import { LoginResponseDto } from './dto/login-response.dto';
-import { User } from './entities/user.entity';
-import { CreateUserDto } from './dto/create-user.dto';
-import { JwtPayload } from './interfaces/jwt-payload.interface';
-
-@Injectable()
-export class AuthService {
-  private readonly logger = new Logger('AuthService');
-
-  constructor(
-    @InjectRepository(User)
-    private readonly userRepository: Repository<User>,
-    private readonly jwtService: JwtService,
-  ) {}
-
-  async create(createUserDto: CreateUserDto) {
-    try {
-      const user = this.userRepository.create(createUserDto);
-      user.activated = true;
-      user.activatedAt = new Date();
-
-      await this.userRepository.save(user);
-      delete user.password;
-      return user;
-    } catch (error) {
-      this.handleDBErrors(error);
-    }
-  }
-
-  async login(loginUserDto: LoginUserDto): Promise<LoginResponseDto> {
-    const { email, password } = loginUserDto;
-
-    const user = await this.userRepository.findOne({
-      where: { email },
-      select: ['email', 'password', 'id'],
-    });
-
-    if (!user) {
-      throw new UnauthorizedException('Credentials are invalid (email)');
-    }
-    if (!bcrypt.compareSync(password, user.password)) {
-      throw new UnauthorizedException('Credentials are invalid (password)');
-    }
-    delete user.email;
-    delete user.password;
-    return {
-      ...user,
-      token: this.getJwtToken({ id: user.id }),
-    };
-  }
-
-  private getJwtToken(payload: JwtPayload) {
-    const token = this.jwtService.sign(payload);
-    return token;
-  }
-
-  private handleDBErrors(error: any): never {
-    if (error.code === '23505') {
-      throw new BadRequestException('Email already exists');
-    }
-
-    throw new InternalServerErrorException();
-  }
-}
