+++ src/auth/commands/handlers/index.ts
@@ -1,0 +1,2 @@
+export { CreateUserHandler } from './create-user.handler';
+export { LoginUserHandler } from './login-user.handler';
