+++ src/auth/commands/impl/create-user.command.ts
@@ -1,0 +1,5 @@
+import { CreateUserDto } from '../../dto';
+
+export class CreateUserCommand {
+  constructor(public readonly createUserDto: CreateUserDto) {}
+}
