+++ src/auth/commands/impl/login-user.command.ts
@@ -1,0 +1,5 @@
+import { LoginUserDto } from './../../dto';
+
+export class LoginUserCommand {
+  constructor(public readonly loginUserDto: LoginUserDto) {}
+}
